Rename ns.js to symbols.js and all the var names
