+++ src/contextTypes.js
@@ -1,10 +1,10 @@
 import {object} from 'prop-types'
-import * as ns from './ns'
+import * as symbols from './symbols'
 import propTypes from './propTypes'
 
 export default {
-  [ns.jss]: propTypes.jss,
-  [ns.sheetOptions]: object,
-  [ns.sheetsRegistry]: propTypes.registry,
-  [ns.managers]: object
+  [symbols.jss]: propTypes.jss,
+  [symbols.sheetOptions]: object,
+  [symbols.sheetsRegistry]: propTypes.registry,
+  [symbols.managers]: object
 }
