+++ src/symbols.js
